% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_article.R
\name{paper_article}
\alias{paper_article}
\title{Title}
\usage{
paper_article(title, meta, text_lead, text, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

if (interactive()) {

ui <- fluidPage(
 paperize(),

 paper_article(
   title = "Introduction to paperrr",
   meta = "written by Mohamed El Fodil Ihaddaden",
   text_lead = "the package relies on the awesome paperCSS library.
                Lorem ipsum dolor sit, amet consectetur adipisicing elit.
                Quos, laborum commodi eaque vitae ad, labore minus necessitatibus,
                fugiat et nemo in impedit! Facilis harum dolorem voluptates,
                voluptatibus laboriosam facere aut?",
   text = "Lorem ipsum, dolor sit amet consectetur adipisicing elit.
           Debitis nobis quisquam enim amet explicabo maxime,
           quasi temporibus fugit, ducimus velit animi voluptas,
           autem repellendus optio!
           Officiis fugiat voluptas cum blanditiis!
           Lorem ipsum dolor sit amet consectetur adipisicing elit.
           Quis eius quidem dignissimos. Quod ratione officiis fugit,
           provident, assumenda ullam aliquid fuga aspernatur sapiente excepturi,
           consectetur corporis qui voluptatem nisi adipisci."
 )
)

server <- function(input, output) {}

shinyApp(ui = ui, server = server)
}

}
