% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_badge.R
\name{paper_badge}
\alias{paper_badge}
\title{Creates a badges}
\usage{
paper_badge(text, type = NULL)
}
\arguments{
\item{text}{the text that will be displayed inside the badge}

\item{type}{the type}
}
\value{

}
\description{
Creates a badges
}
