% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_button.R
\name{paper_button}
\alias{paper_button}
\title{Creates a Shiny actionButton styled with paperCSS}
\usage{
paper_button(
  inputId,
  label,
  size = "default",
  fill = NULL,
  outline_col = NULL,
  icon = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{inputId}{the input ID that will be used to access the butfon}

\item{label}{the label of the button}

\item{size}{the size of the button, either the 'default', 'small',
'large', 'block'. Defaults to 'default'.}

\item{fill}{the color that will be used to fill the button}

\item{outline_col}{the color of the outline of the button}

\item{icon}{an optional icon() to appear on the button}

\item{width}{the width of the button}

\item{...}{names attributes to be applied to the button}
}
\value{
a Shiny button styled with paperCSS
}
\description{
Creates a Shiny actionButton styled with paperCSS
}
