% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_card.R
\name{paper_card}
\alias{paper_card}
\title{Title}
\usage{
paper_card(img_src, img_alt, title, subtitle, text, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

if (interactive()) {
ui <- fluidPage(
  paperize(),
  paper_card(
    img_src = "https://upload.wikimedia.org/wikipedia/commons/thumb/1/1b/R_logo.svg/724px-R_logo.svg.png",
    img_alt = "R logo",
    title = "The R Programming language",
    subtitle = "How Wikpedia defines R",
    text = "R is a programming language and free software environment for statistical
            computing and graphics supported by the R Foundation for
            Statistical Computing. The R language is widely used among
            statisticians and data miners for developing statistical software and data analysis."
  )
)

server <- function(input, output) {}

shinyApp(ui = ui, server = server)
}
}
