% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_card_plain.R
\name{paper_card_plain}
\alias{paper_card_plain}
\title{Title}
\usage{
paper_card_plain(title, subtitle, text, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
if (interactive()) {
ui <- fluidPage(
  paperize(),
  paper_card_plain(
    title = "The R Programming language",
    subtitle = tags$a(href = "https://google.com", "click here"),
    text = "R is a programming language and free software environment for statistical
            computing and graphics supported by the R Foundation for
            Statistical Computing. The R language is widely used among
            statisticians and data miners for developing statistical software and data analysis.",
    h1(a(href = "https://google.com", "click here")),

  )
)

server <- function(input, output) {}
shinyApp(ui = ui, server = server)
}
}
