% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_progress.R
\name{paper_progress}
\alias{paper_progress}
\title{Title}
\usage{
paper_progress(
  progression,
  label = paste0(progression, "\%"),
  style = "primary",
  striped = TRUE
)
}
\arguments{
\item{style}{}
}
\value{

}
\description{
Title
}
\examples{
if (interactive()) {
ui <- fluidPage(
  paperize(),
  paper_progress(progression = 50),
  paper_progress(progression = 50, style = "success"),
  paper_progress(progression = 50, style = "danger"),
  paper_progress(progression = 50, style = "secondary")
)

server <- function(input, output) {}

shinyApp(ui = ui, server = server)
}
}
