% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperize.R
\name{paperize}
\alias{paperize}
\title{Apply the paperCSS styling}
\usage{
paperize()
}
\value{
called for the side effect of applying the
paperCSS styling using a CDN
}
\description{
the function must be called only once,
preferably at the beginning of the Shiny app.
}
\examples{

if (interactive()) {

ui <- fluidPage(

    paperize(),

    h1("Hello paperrr")
)

server <- function(input, output) {}

shinyApp(ui, server)
}
}
